[BH-740] Split SystemManager - part III
Move Pure-specific functionality to Pure SystemManager - part I
