+++ products/PurePhone/services/evtmgr/include/evtmgr/EVMessages.hpp
@@ -1,0 +1,10 @@
+// Copyright (c) 2017-2021, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
+#pragma once
+
+namespace sevm
+{
+    class RequestPhoneModeForceUpdate : public sys::DataMessage
+    {};
+} // namespace sevm
